Divide index.js in modules
