+++ src/modules/getObj.js
@@ -1,0 +1,12 @@
+const url = 'https://www.themealdb.com/api/json/v1/1/filter.php?a=Canadian'
+
+
+const getObj = new Promise((resolve) => {
+    fetch(url)
+    .then(res => res.json())
+    .then(data => {
+       resolve(data.meals.slice(7))
+    })
+ })
+
+ export default getObj;
